Update ep_gui.en-US.rc
Good morning

I have been using ExplorerPatcher for more than two years on a small fleet of PCs and I know how it works quite well.
First of all, I would like to thank you for the excellent work done which allowed me to regain control of the management of my various PCs.
I have been working almost from the beginning on translating the software into French and I did it via the settings.reg and settings10.reg files locally on my PCs to create my .dlls derived from the official .dlls.
I have just noticed with joy that the software is finally available in a multilingual version and therefore I wanted to make my contribution both for the French version and for the native English version because the purpose of the open source software is good to be able to share points of view.
I added a few capital letters or plurals as well as a few hyphens just to catch the eye.
I added a few words such as "simplified" or "Win_Key" or "(S_W_S)" or "switcher" or "proportion" or even "ExplorerPatcher" for the same reasons.
As for the expression "Remember the last used section in this window" I propose this version "Open ExplorerPatcher on the tab of the last used section" which is far from your version but which has the advantage of being better understood by users ordinary people who discover the software.
The same goes for the three texts on the choice of update notifications.

Do not hesitate to give me your feedback on my proposal because I remain open and it is by confronting divergences that we can address relevance.

Very sincerely
Dennis

PS:
Finally, I would like to make a proposal for interactive active display of options in the Windows Switcher tab and in the Weather tab but this action involves modifying the Settings.reg and Setting10.reg files and will therefore be subject to modification. next pull request.