Start10: Prevent menu closing delay when patching standalone (without ExplorerPatcher injecting `explorer.exe`)
The Windows 10 Start menu has a delay when closing if one is injecting
it without injecting `explorer.exe` as well. The reason for this is
that animations are fixed, which in turn fixes this delay, using code
run when ExplorerPatcher is injected in `explorer.exe`. It cannot be
patched from the Start menu's host process itself
(`StartMenuExperienceHost.exe`) because that doesn't load
`twinui.pchsell.dll` where the actual patching happens. This commit
works around this shortcoming by employing the old fix for this delay:
hiding the Start menu when animations haven't been fixed (so when
`explorer.exe` is not injected by ExplorerPatcher).

We needed to detect when animations were successfully patched from
`explorer.exe` in order to disable this workaround, thus this commit
offers an example on how to create a kernel object, in this case a
mutex, that can be 'seen' by a process running inside an app container,
as is the case for Start menu's host process
(`StartMenuExperienceHost.exe`).